iniciando component dashboard
