+++ backend/src/server.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const mongoose = require("mongoose");
+const cors = require("cors");
 
 const routes = require("./routes");
 
@@ -19,8 +20,8 @@ mongoose
     console.log("error ao conectar no banco " + err);
   });
 
+app.use(cors());
 app.use(express.json());
-
 app.use(routes);
 
 const port = 3001;
