+++ frontend/package.json
@@ -6,6 +6,7 @@
     "axios": "^0.19.0",
     "react": "^16.10.1",
     "react-dom": "^16.10.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.1.2"
   },
   "scripts": {
