+++ frontend/src/pages/Dashboard/Index.js
@@ -1,0 +1,7 @@
+import React, { useEffect } from "react";
+
+export default function Dashboard() {
+  useEffect();
+
+  return <h1>Dashboard</h1>;
+}
