+++ frontend/src/pages/New/Index.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function New() {
+  return <h1>New</h1>;
+}
